ignore non-directory entries in manifests array
The manifests array should contain valid path entries, however, it might
be possible, that sometimes those directories don't exist or are not a
directory at all. In these cases, the entries should just be ignored,
without triggering an error message. In verbose mode, a warning is
printed to the console.