hotfix replace_version for proxy mode for node 0.12
the proxy module needed to recompress the stream and transform it, in
order to replace the version string for cache busting reasons during
local development. This eventually has an impact on the content length
of the HTTP response, which might have confused browsers. In case the
stream is compressed, we just decompress it and remove the
content-encoding headers, to work around this issue. So the transformed
files will never be sent compressed from appserver to the browser. Which
is a reasonable trade off.

This might be fixed in a better way, in the future, but this would need
major refactoring of the middleware modules. At the moment, the code is
just not flexible enough to handle such cases in a maintainable way.