fix problem with binary data being proxied
most binary data happened to be around twice in size, when piped through
the proxy middlerware. This was, because of node trying to encode the
response using UTF-8, when in fact the encoding should have been
'binary'. Setting up the proxy using readableStream.pipe is obviously
not the way to go when piping binary data through a stream.