write correct headers from the actual login response
some implementations of the login API might actually contain useful
information in the header, like cookies. So we just forward the headers
from the login proxy request, so the client can decide what to do.

Fixes issues with login servlets that send cookies needed for further
communication with the backend.