forward HTTP headers to apps/load remote endpoint
the original request might contain cookies and other headers that are
needed to complete the request on the backend. So all headers from the
browser are now forwarded to the remote endpoint when fetching files via
appsload. All but the Accept-Encoding header that is, because the
content is supposed to be transformed and should not be compressed.