bump version
we want to release, what has been 0.6.2 as 0.7.0 (final) because we
published pre-releases to npm but don't want to release this version
under this name any longer. So this helps us to deprecate the old npm
module in favour of the new one which is going to be released as soon as
it's ready.